+++ frontend/src/components/wms/execution/index.ts
@@ -3,11 +3,24 @@ export { ScanStepHero } from "./ScanStepHero";
 export { ExecutionBottomBar, EXECUTION_BOTTOM_RESERVE } from "./ExecutionBottomBar";
 export { ExecutionTouchButton } from "./ExecutionTouchButton";
 export { ExecutionGlobalContextBar } from "./ExecutionGlobalContextBar";
+export { ActiveOperationContextBar } from "./ActiveOperationContextBar";
+export {
+  ACTIVE_OPERATION_CONTEXT_BAR_OFFSET,
+  normalizeOperationContext,
+  formatOrderNumberLabel,
+  formatCartLabel,
+  formatOperatorDisplayName,
+} from "./activeOperationContext";
 export { ScanFeedbackOverlay } from "./ScanFeedbackOverlay";
 export { useWmsPageScanHandler } from "./useWmsPageScanHandler";
 export { useWmsScanResolveNavigate } from "./useWmsScanResolveNavigate";
 export { useScanFeedback } from "./useScanFeedback";
 export { useOfflineActionQueue } from "./useOfflineActionQueue";
 export { formatOperationalError } from "./formatOperationalError";
-export { executionContextFromOperationalDetail } from "./syncExecutionContext";
+export {
+  executionContextFromOperationalDetail,
+  executionContextFromPicking,
+  executionContextFromPacking,
+  executionContextFromPutaway,
+} from "./syncExecutionContext";
 export { isWarehouseExecutionRoute } from "./executionRoutes";